requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
future==0.18.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras-Applications==1.0.8
Markdown==3.2.1
matplotlib
mkl-fft
mkl-random
mkl-service
numpy
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.1
Pillow
pipreqs==0.4.10
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing
python-dateutil
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six
termcolor==1.1.0
#torch==1.7.1
#torchvision
tornadotorchvision==0.8.2
typing-extensions
udacity-pa==0.2.9
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
